CYBERSHIELD/cybershield/backend/requirements.txt
Flask
flask-cors
flask-bcrypt
flask_jwt_extended
scikit-learn
numpy
pandas
joblib
matplotlib
seaborn
python-dotenv
pymongo
spacy
nltk
bs4
gunicorn